+++ Test/BlankTest.ps1
@@ -5,7 +5,7 @@ MIT License
 This file is part of "Windows Firewall Ruleset" project
 Homepage: https://github.com/metablaster/WindowsFirewallRuleset
 
-Copyright (C) 2020-2022 metablaster zebal@protonmail.ch
+Copyright (C) 2022 metablaster zebal@protonmail.ch
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -26,84 +26,48 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 #>
 
-# Firewall rules exist for programs listed here
-# Services and store apps are not part of this enumeration
-enum TargetProgram
-{
-	MSI
-	EdgeWebView
-	dotnet
-	CMake
-	SqlPath
-	SqlServer
-	SqlManagementStudio
-	WindowsDefender
-	NuGet
-	NETFramework
-	vcpkg
-	SysInternals
-	WindowsKits
-	WebPlatform
-	XTU
-	Chocolatey
-	ArenaChess
-	GoogleDrive
-	RivaTuner
-	Incredibuild
-	ColorMania
-	MetaTrader
-	RealWorld
-	AzureDataStudio
-	qBittorrent
-	OpenTTD
-	EveOnline
-	DemiseOfNations
-	CounterStrikeGO
-	PinballArcade
-	JavaUpdate
-	JavaRuntime
-	AdobeARM
-	AdobeReader
-	AdobeAcrobat
-	LoLGame
-	FileZilla
-	PathOfExile
-	HWMonitor
-	CPUZ
-	MSIAfterburner
-	GPG
-	OBSStudio
-	PasswordSafe
-	Greenshot
-	DnsCrypt
-	OpenSSH
-	PowerShellCore64
-	PowerShell64
-	PowerShell86
-	OneDrive
-	HelpViewer
-	VSCode
-	MicrosoftOffice
-	TeamViewer
-	EdgeChromium
-	Chrome
-	Firefox
-	Yandex
-	Tor
-	uTorrent
-	Thuderbird
-	Steam
-	Nvidia64
-	Nvidia86
-	GeForceExperience
-	WarThunder
-	PokerStars
-	VisualStudio
-	VisualStudioInstaller
-	MSYS2
-	Git
-	GitHubDesktop
-	EpicGames
-	UnrealEngine
-	BingWallpaper
-}
+<#
+.SYNOPSIS
+Empty unit test
+
+.DESCRIPTION
+Use to temporarily test whatever
+
+.PARAMETER Force
+If specified, this unit test runs without prompt to allow execute
+
+.EXAMPLE
+PS> .\BlankTest.ps1
+
+.INPUTS
+None. You cannot pipe objects to BlankTest.ps1.
+
+.OUTPUTS
+None. BlankTest.ps1 does not generate any output.
+
+.NOTES
+None.
+#>
+
+#Requires -Version 5.1
+#Requires -RunAsAdministrator
+
+[CmdletBinding()]
+param (
+	[Parameter()]
+	[switch] $Force
+)
+
+#region Initialization
+. $PSScriptRoot\..\Config\ProjectSettings.ps1 $PSCmdlet
+. $PSScriptRoot\ContextSetup.ps1
+
+Initialize-Project -Strict
+if (!(Approve-Execute -Accept $Accept -Deny $Deny -Force:$Force)) { exit }
+#endregion
+
+Enter-Test
+Start-Test "Blank test"
+
+Update-Log
+Exit-Test